# konto-verwaltung-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Minh Chi Diep",
    "description": "OpenApi documentation for account management",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Minh Chi Diep",
      "url": "https://minhchi1709.github.io/profile/",
      "email": "diepminhchi1617@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Local ENV"
    }
  ],
  "paths": {
    "/app/AccountManagement/api/v1/transactions/transactions/{id}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getTransactionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Transaction"
        ],
        "operationId": "updateTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transaction"
        ],
        "operationId": "deleteTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllYearTotals",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/YearTotal"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transaction"
        ],
        "operationId": "createTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/currency": {
      "get": {
        "tags": [
          "Currency"
        ],
        "operationId": "getAllRates",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rates"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Currency"
        ],
        "operationId": "changeRate",
        "parameters": [
          {
            "name": "bank",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllYears",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years/{year}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllTransactionsOfYear",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years/{year}/months": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllMonths",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "JANUARY",
                      "FEBRUARY",
                      "MARCH",
                      "APRIL",
                      "MAY",
                      "JUNE",
                      "JULY",
                      "AUGUST",
                      "SEPTEMBER",
                      "OCTOBER",
                      "NOVEMBER",
                      "DECEMBER"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years/{year}/months/{month}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllTransactionsOfMonth",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years/{year}/month-total": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllMonthTotals",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonthTotal"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/savings": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllSavings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/salaries": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllSalaries",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/currency/today": {
      "get": {
        "tags": [
          "Currency"
        ],
        "operationId": "getTodayRate",
        "parameters": [
          {
            "name": "bank",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyExchangeRate"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TransactionRequest": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "value": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "YearTotal": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "MonthTotal": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "enum": [
              "JANUARY",
              "FEBRUARY",
              "MARCH",
              "APRIL",
              "MAY",
              "JUNE",
              "JULY",
              "AUGUST",
              "SEPTEMBER",
              "OCTOBER",
              "NOVEMBER",
              "DECEMBER"
            ]
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CurrencyExchangeRate": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/CurrencyExchangeRateId"
          },
          "rate": {
            "type": "number",
            "format": "float"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CurrencyExchangeRateId": {
        "type": "object",
        "properties": {
          "bank": {
            "type": "string",
            "enum": [
              "VIB",
              "VCB",
              "PAYPAL"
            ]
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "Rates": {
        "type": "object",
        "properties": {
          "paypalRates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurrencyExchangeRate"
            }
          },
          "vibRates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurrencyExchangeRate"
            }
          },
          "vcbRates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurrencyExchangeRate"
            }
          }
        }
      }
    }
  }
}